requirements.txt
flask
tensorflow
kafka-python
numpy
scikit-learn
